cheri: Make drivers/scsi/ compile and fix warnings
- Standard uintptr_t vs. unsigned long issues.
- In drivers/arcmsr/ we must be careful to retain provenance when
  calculating virtual from physical addresses via an offset.
- drivers/scsi/sym53c8xx_2/sym_malloc.c contains a hand crafred
  buddy allocator that does not set bounds.
- In drivers/scsi/esas2r/ there is one ioctl that provides kernel
  pointers to user space presumably for debugging purposes. To
  preserv the ioctl ABI for the driver convert the void pointers
  in the ioctl structure to unsigned longs.

Driver that we ignore for the time being:
- CONFIG_SCSI_BFA_FC and some of CONFIG_MEGARAID_* is complicated
  and postponed. Will be disabled in the default config.
- CONFIG_SCSI_SNIC and CONFIG_SCSI_CHELSIO_FCOE uses unsigned long
  callback data in hardware buffers that will not work with CHERI.
- drivers/scsi/aacraid/ has a user space API that splits 64-bit
  addresses into 32-bit chunks. With the userland tools and the
  hardware available for testing this could be fixed easily.

+++ drivers/scsi/elx/efct/efct_hw.h
@@ -628,7 +628,7 @@ efct_hw_io_init_sges(struct efct_hw *hw,
 
 int
 efct_hw_io_add_sge(struct efct_hw *hw, struct efct_hw_io *io,
-		   uintptr_t addr, u32 length);
+		   unsigned long addr, u32 length);
 int
 efct_hw_io_abort(struct efct_hw *hw, struct efct_hw_io *io_to_abort,
 		 bool send_abts, void *cb, void *arg);
